requirements.txt
anthropic
python-dotenv
langchain
langchain-anthropic
langchain-openai
chromadb
crewai
streamlit
openai
tzdata
